+++ src/features/blog/components/RelatedPosts.tsx
@@ -1,4 +1,5 @@
-import type { PostSummary } from "../types";
+import type { PostSummary } from "@/features/blog/types";
+
 import { BlogPostCard } from "./BlogPostCard";
 
 type RelatedPostsProps = {
