+++ src/features/tags/services/api.ts
@@ -1,6 +1,5 @@
 import { getAllPosts } from "@/features/blog";
-
-import type { TagCount } from "../types";
+import type { TagCount } from "@/features/tags/types";
 
 /**
  * 모든 태그와 포스트 개수 가져오기
