+++ src/shared/components/mdx/index.ts
@@ -1,7 +1,7 @@
 export { CustomMDX } from "./CustomMDX";
 export { MdxCode } from "./MdxCode";
 export { createHeading } from "./MdxHeading";
-export { MdxImage } from "./MdxImage";
+export { MdxImage, MdxImg } from "./MdxImage";
 export { MdxLink } from "./MdxLink";
 export { MdxPre } from "./MdxPre";
 export { MdxTable, MdxTbody, MdxTd, MdxTh, MdxThead, MdxTr } from "./MdxTable";
